adapts Gobra to renamed ViperServer packages
