+++ src/main/scala/viper/gobra/backend/ViperVerifier.scala
@@ -1,6 +1,6 @@
 package viper.gobra.backend
 
-import viper.server.ViperBackendConfig
+import viper.server.core.ViperBackendConfig
 import viper.silver
 import viper.silver.reporter.Reporter
